xiaozhi-proxy/requirements.txt
aiohttp>=3.8.0
aiomqtt>=2.0.0
websockets>=10.0
pycryptodome
pyyaml>=6.0
